[BRO Assistant] Guard My Gang - BRS-0018: Clean-up code
